requirements.txt
asgiref==3.5.2
beautifulsoup4==4.10.0
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.4
Django==4.0.5
filelock==3.7.0
google-api-core==1.28.0
google-api-python-client==2.5.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
lxml==4.8.0
oauthlib==3.1.0
packaging==20.9
platformdirs==2.5.2
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-decouple==3.6
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
tzdata==2022.1
uritemplate==3.0.1
urllib3==1.26.4
virtualenv==20.14.1
whitenoise==6.2.0
psycopg2==2.9.3
